fix: rename exported proposal view module for naming consistency
