+++ src/front-end/typescript/lib/pages/proposal/team-with-us/lib/views/exported-proposal.tsx
